requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# NLP and embeddings
gensim>=4.2.0
transformers>=4.20.0
torch>=1.12.0
datasets>=2.0.0
tokenizers>=0.12.0

# Web scraping and APIs
requests>=2.28.0
beautifulsoup4>=4.11.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
psutil>=5.9.0

# Optional: for better performance
# accelerate>=0.20.0
# optimum>=1.8.0
